+++ semcore/data-table/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Row hover effect for empty state.
 - Incorrect 'Select all' checkbox behavior with pagination.
+- Header glitches in loading state with sticky header.
 
 ## [16.0.6] - 2025-07-04